+++ src/Server/Common/Game.cs
@@ -50,15 +50,15 @@ namespace LambdaRobots.Server {
         public List<Message> Messages { get; set; } = new List<Message>();
 
         // game characteristics
-        public double BoardWidth { get; set; }
-        public double BoardHeight { get; set; }
-        public double SecondsPerTurn { get; set; }
-        public double DirectHitRange { get; set; }
-        public double NearHitRange { get; set; }
-        public double FarHitRange { get; set; }
-        public double CollisionRange { get; set; }
-        public double MinRobotStartDistance { get; set; }
-        public double RobotTimeoutSeconds { get; set; }
+        public float BoardWidth { get; set; }
+        public float BoardHeight { get; set; }
+        public float SecondsPerTurn { get; set; }
+        public float DirectHitRange { get; set; }
+        public float NearHitRange { get; set; }
+        public float FarHitRange { get; set; }
+        public float CollisionRange { get; set; }
+        public float MinRobotStartDistance { get; set; }
+        public float RobotTimeoutSeconds { get; set; }
         public int MaxTurns { get; set; }
         public int MaxBuildPoints { get; set; }
     }
@@ -96,32 +96,32 @@ namespace LambdaRobots.Server {
         public MissileStatus Status { get; set; }
 
         [DataMember(IsRequired = true)]
-        public double X { get; set; }
+        public float X { get; set; }
 
         [DataMember(IsRequired = true)]
-        public double Y { get; set; }
+        public float Y { get; set; }
 
         [DataMember(IsRequired = true)]
-        public double Distance { get; set; }
+        public float Distance { get; set; }
 
         // missile characteristics
 
         [DataMember(IsRequired = true)]
-        public double Speed { get; set; }
+        public float Speed { get; set; }
 
         [DataMember(IsRequired = true)]
-        public double Heading { get; set; }
+        public float Heading { get; set; }
 
         [DataMember(IsRequired = true)]
-        public double Range { get; set; }
+        public float Range { get; set; }
 
         [DataMember(IsRequired = true)]
-        public double DirectHitDamageBonus { get; set; }
+        public float DirectHitDamageBonus { get; set; }
 
         [DataMember(IsRequired = true)]
-        public double NearHitDamageBonus { get; set; }
+        public float NearHitDamageBonus { get; set; }
 
         [DataMember(IsRequired = true)]
-        public double FarHitDamageBonus { get; set; }
+        public float FarHitDamageBonus { get; set; }
     }
 }
